requirements.txt
gradio>=4.0.0
anthropic>=0.18.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
python-dotenv>=1.0.0
numpy

